Don't split sending CommitText when auto-start conversion
When a candidate is selected as a result of typing an auto-start
conversion character (e.g., "、"), CommitText signal is sent twice:
one for the candidate and the other for the trigger character, which
may confuse IBus and the former CommitText may be ignored.  To work
around the behavior, this patch coalesce those CommitText signals,
while still sending it separately when the candidate is clicked by
mouse.

Signed-off-by: Daiki Ueno <dueno@redhat.com>